Split folder UI into separate screen-tab-bar-folders feature
Moves FLDR button + XIT FLDR registration out of screen-tab-bar into a
new screen-tab-bar-folders feature so users can disable folders
independently via userData.settings.disabled.
