+++ src/features/basic/screen-tab-bar/screen-tab-bar.tsx
@@ -1,7 +1,5 @@
 import $style from './screen-tab-bar.module.css';
 import TabBar from './TabBar.vue';
-import FolderCreator from './FolderCreator.vue';
-import FldrButton from './FldrButton.vue';
 import { userData } from '@src/store/user-data';
 import removeArrayElement from '@src/utils/remove-array-element';
 import { watchEffectWhileNodeAlive } from '@src/utils/watch';
@@ -72,18 +70,6 @@ function extractScreenId(url?: string) {
 
 function onContainerReady(container: HTMLElement) {
   createFragmentApp(TabBar).appendTo(container);
-  let fldrInserted = false;
-  subscribe($$(container, C.HeadItem.label), label => {
-    if (fldrInserted || label.textContent !== 'FULL') {
-      return;
-    }
-    const fullItem = label.closest(`.${C.HeadItem.container}`) as HTMLElement | null;
-    if (!fullItem) {
-      return;
-    }
-    fldrInserted = true;
-    createFragmentApp(FldrButton).before(fullItem);
-  });
 }
 
 function init() {
@@ -91,13 +77,6 @@ function init() {
   subscribe($$(document, C.ScreenControls.screens), onListReady);
   applyCssRule(`.${C.Head.contextAndScreens}`, $style.contextAndScreens);
   applyCssRule(`.${C.ScreenControls.container}`, $style.screenControls);
-  xit.add({
-    command: 'FLDR',
-    name: 'NEW FOLDER',
-    description: 'Create a new screen folder.',
-    component: () => FolderCreator,
-    bufferSize: [450, 110],
-  });
 }
 
 features.add(import.meta.url, init, 'Adds a tab bar for user screens.');
